configurable_NIAnalogIn: Added support for configurable voltage range on NI analog inputs.
A new AnalogIn class has been created in NIBoard.py called NIAnalogIn. The signature of this class is backwards compatible with AnalogIn, and the defaults for the new keyword arguments (min/max) match those hard coded into the existing NI BLACS classes.
The NI classes with analog inputs have been updated to use the min/max values specified in the device properties of the NI card.

All changes are backwards compatible with the previous API and should maintain the original behaviour if no modifications are made to the labscript file.